requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
nltk==3.8.1
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
pydantic==2.5.2
typing-extensions==4.8.0
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
soupsieve==2.5
tqdm==4.66.1
regex==2023.10.3 